# Move logic of bucket avaliation (addSort, addWild...) from sortTriggers to Inheritance class # Add some método do fill the bucket with unsorted list # Add some static helper to do FIll and Dump at same time (SortAtOnce) # Loop on prev2trig dictionary
I think it's not the most performatic way for this, but it's not so bad =]
